fix: Remove SafeCast in Pool and add use of uint128 for flashloan premiums

+++ contracts/interfaces/IPool.sol
@@ -631,8 +631,8 @@ interface IPool {
    * @param flashLoanPremiumToProtocol The part of the premium sent to the protocol treasury, expressed in bps
    */
   function updateFlashloanPremiums(
-    uint256 flashLoanPremiumTotal,
-    uint256 flashLoanPremiumToProtocol
+    uint128 flashLoanPremiumTotal,
+    uint128 flashLoanPremiumToProtocol
   ) external;
 
   /**
@@ -681,7 +681,7 @@ interface IPool {
    * @notice Returns the total fee on flash loans
    * @return The total fee on flashloans
    */
-  function FLASHLOAN_PREMIUM_TOTAL() external view returns (uint256);
+  function FLASHLOAN_PREMIUM_TOTAL() external view returns (uint128);
 
   /**
    * @notice Returns the part of the bridge fees sent to protocol
@@ -693,7 +693,7 @@ interface IPool {
    * @notice Returns the part of the flashloan fees sent to protocol
    * @return The flashloan fee sent to the protocol treasury
    */
-  function FLASHLOAN_PREMIUM_TO_PROTOCOL() external view returns (uint256);
+  function FLASHLOAN_PREMIUM_TO_PROTOCOL() external view returns (uint128);
 
   /**
    * @notice Returns the maximum number of reserves supported to be listed in this Pool
